Secure user management endpoints

+++ servicex_app/servicex/resources/jwt/pending_all.py
@@ -25,14 +25,25 @@
 # CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-from flask_restful import Resource
+from flask_jwt_extended import jwt_optional
 
+from resources.servicex_resource import ServiceXResource
 from servicex.models import UserModel
 
 
-class Pending_AllUsers(Resource):
+class PendingAllUsers(ServiceXResource):
+    @jwt_optional
     def get(self):
-        return UserModel.return_all_pending()
+        is_auth, auth_reject_message = self._validate_user_is_admin()
+        if not is_auth:
+            return {'message': f'Authentication Failed: {str(auth_reject_message)}'}, 401
+        else:
+            return UserModel.return_all_pending()
 
+    @jwt_optional
     def delete(self):
-        return UserModel.delete_all_pending()
+        is_auth, auth_reject_message = self._validate_user_is_admin()
+        if not is_auth:
+            return {'message': f'Authentication Failed: {str(auth_reject_message)}'}, 401
+        else:
+            return UserModel.delete_all_pending()
